refactor: extract CheckForKnownSkipConditions to separate method
